V0.0.0.4

+++ src/app.update.ts
@@ -11,7 +11,6 @@ import { PinoLoggerService } from './loger/pino.loger.service';
 import { MyMessage } from './interface/my-message.interface';
 import { TransactionType } from './mongodb/shemas';
 import { Context, CustomCallbackQuery } from './interface/context.interfsce';
-import { callbackQuery } from 'telegraf/filters';
 
 @Update()
 export class AppUpdate {
@@ -26,7 +25,7 @@ export class AppUpdate {
   @Start()
   async startCommand(ctx: Context) {
     try {
-      const userId = ctx.message.from.id;
+      const userId = ctx.from.id;
       await this.transactionService.createBalance({ userId });
       await ctx.reply(
         '\n' +
@@ -43,9 +42,7 @@ export class AppUpdate {
       this.logger.log('startCommand executed successfully');
     } catch (error) {
       this.logger.error('Error in startCommand:', error);
-      await ctx.reply(
-        'Произошла ошибка при выполнении команды. Пожалуйста, попробуйте еще раз позже.',
-      );
+      await ctx.reply('Произошла ошибка при выполнении команды. Пожалуйста, попробуйте еще раз позже.');
     }
   }
   @Hears('Транзакция 💸')
@@ -58,14 +55,14 @@ export class AppUpdate {
   @Action('Приход')
   async incomeCommand(ctx: Context) {
     ctx.session.type = 'income';
-    await ctx.reply('Введите наименование прихода и сумму через пробел');
+    await ctx.reply('Введите наименование прихода и сумму через пробел' + '\n' + 'Пример: "Зарплата 100000"');
     this.logger.log('Приход command executed');
     await ctx.deleteMessage();
   }
   @Action('Расход')
   async expenseCommand(ctx: Context) {
     ctx.session.type = 'expense';
-    await ctx.reply('Введите наименование расхода и сумму через пробел');
+    await ctx.reply('Введите наименование расхода и сумму через пробел' + '\n' + 'Пример: "Продукты 1000"');
     this.logger.log('Расход command executed');
     await ctx.deleteMessage();
   }
@@ -73,62 +70,34 @@ export class AppUpdate {
   async incomeListCommand(ctx: Context) {
     this.logger.log('приходы command executed');
     const userId = ctx.from.id;
-    await this.transactionService.getTransactionsByType(
-      userId,
-      TransactionType.INCOME,
-    );
+    await this.transactionService.getTransactionsByType(userId, TransactionType.INCOME);
     this.logger.log('приходы command executed');
   }
   @Action('Мои расходы')
   async expenseListCommand(ctx: Context) {
     this.logger.log('расходы command executed');
     const userId = ctx.from.id;
-    await this.transactionService.getTransactionsByType(
-      userId,
-      TransactionType.EXPENSE,
-    );
+    await this.transactionService.getTransactionsByType(userId, TransactionType.EXPENSE);
     this.logger.log('расходы command executed');
   }
   @Action('По категории')
   async categoryListCommand(ctx: Context) {
     const userId = ctx.from.id;
-    const uniqueTransactionNames =
-      await this.transactionService.getUniqueTransactionNames(userId);
-    const transactionNameButtons = actionButtonsTransactionNames(
-      uniqueTransactionNames,
-    );
+    const uniqueTransactionNames = await this.transactionService.getUniqueTransactionNames(userId);
+    const transactionNameButtons = actionButtonsTransactionNames(uniqueTransactionNames);
     await ctx.reply('Выберите категорию транзакции:', transactionNameButtons);
   }
 
   @Action(/TransactionName:(.+)/)
   async transactionNameCommand(ctx: Context) {
     this.logger.log('transactionName command executed');
-    const callbackQuery: CustomCallbackQuery =
-      ctx.callbackQuery as CustomCallbackQuery;
+    const callbackQuery: CustomCallbackQuery = ctx.callbackQuery as CustomCallbackQuery;
     if (callbackQuery) {
       const callbackData = callbackQuery.data;
       const parts = callbackData.split(':');
       const selectedTransactionName = parts[1];
       const userId = ctx.from.id;
-      const transactions =
-        await this.transactionService.getTransactionsByTransactionName(
-          userId,
-          selectedTransactionName,
-        );
-      if (transactions.length > 0) {
-        const formattedTransactions = transactions.map(
-          this.transactionService.formatTransaction,
-        );
-        await ctx.reply(
-          `Транзакции по категории "${selectedTransactionName}":\n${formattedTransactions.join(
-            '\n',
-          )}`,
-        );
-      } else {
-        await ctx.reply(
-          `Нет транзакций по категории "${selectedTransactionName}"`,
-        );
-      }
+      await this.transactionService.getTransactionsByTransactionName(userId, selectedTransactionName);
     } else {
       this.logger.log('callbackQuery is undefined');
     }
@@ -163,19 +132,14 @@ export class AppUpdate {
   async listCommand(ctx: Context) {
     try {
       await ctx.deleteMessage();
-      const userId = ctx.message.from.id;
+      const userId = ctx.from.id;
       await this.transactionService.getBalance(userId);
       ctx.session.type = 'balance';
-      await ctx.reply(
-        'Хотите получить статистику транзакций? Выберите параметр: ',
-        actionButtonsStatistics(),
-      );
+      await ctx.reply('Хотите получить статистику транзакций? Выберите параметр: ', actionButtonsStatistics());
       this.logger.log('Баланс command executed');
     } catch (error) {
       this.logger.error('Error in listCommand:', error);
-      await ctx.reply(
-        'Произошла ошибка при выполнении команды. Пожалуйста, попробуйте еще раз позже.',
-      );
+      await ctx.reply('Произошла ошибка при выполнении команды. Пожалуйста, попробуйте еще раз позже.');
     }
   }
   @On('text')
@@ -184,28 +148,43 @@ export class AppUpdate {
       return;
     }
     const message = ctx.message as MyMessage;
-    const userId = ctx.message.from.id;
+    const userId = ctx.from.id;
     const text = message.text;
-    const [transactionName, sum] = text.split(' ');
-    const amount = Number(sum);
-    const transactionType =
-      ctx.session.type === 'income'
-        ? TransactionType.INCOME
-        : TransactionType.EXPENSE;
-    await this.transactionService.createTransaction({
-      userId,
-      transactionName,
-      transactionType,
-      amount,
-    });
-    await this.transactionService.updateBalance(
-      userId,
-      amount,
-      transactionType,
-    );
-    await this.transactionService.getBalance(userId);
-    await ctx.deleteMessage();
-    delete ctx.session.type;
-    this.logger.log('textCommand executed');
+    const regex = /^([a-zA-Zа-яА-Я]+(?:\s+[a-zA-Zа-яА-Я]+)?)\s+([\d.]+)$/;
+    const matches = text.match(regex);
+    if (!matches) {
+      await ctx.reply('Пожалуйста, введите корректные данные.');
+      return;
+    }
+    const transactionName = matches[1].trim().toLowerCase(); // Преобразование в нижний регистр
+    const amount = Number(matches[matches.length - 1]);
+    if (!transactionName || isNaN(amount) || amount <= 0) {
+      await ctx.reply('Пожалуйста, введите корректные данные.');
+      return;
+    }
+    const words = transactionName.split(' ');
+    if (words.length > 2) {
+      await ctx.reply('Пожалуйста, укажите одно или два слова в поле "Наименование транзакции".');
+      return;
+    }
+    const transactionType = ctx.session.type === 'income' ? TransactionType.INCOME : TransactionType.EXPENSE;
+    try {
+      await this.transactionService.createTransaction({
+        userId,
+        transactionName,
+        transactionType,
+        amount,
+      });
+      await this.transactionService.updateBalance(userId, amount, transactionType);
+      await this.transactionService.getBalance(userId);
+
+      await ctx.deleteMessage();
+      delete ctx.session.type;
+      this.logger.log('textCommand executed');
+    } catch (error) {
+      // Обработка ошибок
+      this.logger.error('Error in textCommand:', error);
+      await ctx.reply('Произошла ошибка при выполнении команды. Пожалуйста, попробуйте еще раз позже.');
+    }
   }
 }
